api/requirements.txt
azure-functions>=1.12.0